mui_project/requirements.txt
streamlit
tensorflow
opencv-python
numpy
Pillow
torch
transformers